just saving my changes here

+++ eb-dashboard-backend/src/main/java/com/bcg/ebdashboardbackend/repository/ReviewerRepository.java
@@ -3,5 +3,5 @@ package com.bcg.ebdashboardbackend.repository;
 import com.bcg.ebdashboardbackend.entity.Reviewer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewerRepository  extends JpaRepository<Reviewer, Integer> {
+public interface ReviewerRepository  extends JpaRepository<Reviewer, Long> {
 }
